Add contactid attribute and update entity definitions
Updated `ContactCreatePlugin.model.json` and `Entities.model.json` to include the `contactid` attribute in the `contact` entity. Enhanced metadata with new `MetadataId` values. Introduced a new partial class `Partner` in `Entities.cs` for the `account` entity, along with necessary using directives. Updated `Entities.yaml` to reflect the new structure for the `account` entity. Modified `XrmGenTest.csproj` to include `Entities.yaml` for generating `Entities.cs` and updated the `EnvironmentUrl`. These changes improve entity management and structure.
