+++ src/Testbed/XrmGenTest/Entities.yaml
@@ -1,0 +1,3 @@
+entities:
+- entity: account
+  attributeNames: name
