+++ src/Testbed/XrmGenTest/Model/Entities.cs
@@ -4,7 +4,7 @@ using XrmTools.Meta.Attributes;
 
 [assembly: Entity("notification", AttributeNames = "createdon,eventid,notificationnumber,notificationid")]
 [assembly: Entity("account", AttributeNames = "name,accountnumber,telephone1,statecode,statuscode,primarycontactid,address1_city")]
-[assembly: Entity("contact", AttributeNames = "name,parentcontactid,annualincome,territorycode,accountid,accountidyominame")]
+[assembly: Entity("contact", AttributeNames = "name,parentcontactid,annualincome,territorycode,accountid,accountidyominame,contactid")]
 
 public class Test
 {
